+++ docs/versionhistory.rst
@@ -9,6 +9,10 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
 - Fixed ``__hash__()`` of the C version of the ``CBORTag`` type crashing when there's a recursive
   reference cycle
+- Fixed type annotation for the file object in ``cbor2.dump()``, ``cbor2.load()``, ``CBOREncoder``
+  and ``CBORDecoder`` to be ``IO[bytes]`` instead of ``BytesIO``
+- Changed the return type annotations of ``cbor2.load()`` and ``cbor2.load()`` to return ``Any``
+  instead of ``object`` so as not to force users to make type casts
 - Fixed decoding of epoch-based dates being affected by the local time zone in the C extension
 
 **5.6.1** (2024-02-01)